Move watch mode code into a separate file.
